final projek

+++ database/migrations/2017_06_07_063713_create_dend_table.php
@@ -15,7 +15,7 @@ class CreateDendTable extends Migration
     {
         Schema::create('dendas', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('harga')->unsigned();
+            $table->double('harga')->unsigned();
             $table->timestamps();
         });
     }
